Adds a lot of documentation
