+++ .yardoc/checksums
@@ -1,11 +1,11 @@
-lib/text.rb 48cecb19a69da3b2272bd53efe89a8a42453cc86
-lib/tree.rb 1077fa4705ded68542846bca30b7a9b4349d4c38
-lib/token.rb 61e6f103513aa9a44b847055ccedd9e18f28cffe
-lib/mention.rb 08934f7701bf8efbc57c4cea59d46dddf507d526
-lib/sentence.rb 2b67cd206ecaa9d690160886d3e437396dcd35f8
-lib/coref_chain.rb 8df999257e9dcc0d578283ece3c5c69ed9862a35
-lib/nlp_wrapper.rb 9463e9df559e6243ea76ae591ef510e62f3dd0f1
-lib/nlp_pipeline.rb 2ec8678797b7de1de65633e815a13d3f6a3405e6
-lib/coref_chain_map.rb 930623c235e5cfb46c826f541137bb8c258c05c1
-lib/text_preprocessor.rb e01343ce6f0f22e02f4fe6eb6dd99f598298ab78
+lib/text_preprocessor.rb 52cb529be0ce066fae334566fef6619f717518b1
+lib/stanford_core/text.rb 2798b4352a53d3d9f567df8805e97193b41ad465
+lib/stanford_core/tree.rb f23db1d7ae5a65a19081bd3989bc9e721f73254e
+lib/stanford_core/token.rb 5a4db3b75c5e53f3d8fab845853bfa292be6464e
+lib/stanford_core/mention.rb 1a2a8e4733dfa43381c4047a828d172b5576a799
+lib/stanford_core/sentence.rb fd2c4fd686f822d5cada1c349af86837c8fb6b67
+lib/stanford_core/coref_chain.rb 9f58c700e959461224fffa3280f074d762287052
+lib/stanford_core/nlp_wrapper.rb 0ea3310ba52cf42fc9a8525733b214f8473a8d31
+lib/stanford_core/nlp_pipeline.rb 40cc5bb5c5a0a534fc37300c29a5a9398c8762e1
 lib/tree_activities_identifier.rb 7ed3f75d0a87be0da37ea3aab16ffa43b7a64de2
+lib/stanford_core/coref_chain_map.rb a207238370adb120a0104dcc1ae276d080ff8bad
