+++ doc/StanfordCore.html
@@ -1,0 +1,128 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8">
+<meta name="viewport" content="width=device-width, initial-scale=1.0">
+<title>
+  Module: StanfordCore
+  
+    &mdash; BPM Trello
+  
+</title>
+
+  <link rel="stylesheet" href="css/style.css" type="text/css" />
+
+  <link rel="stylesheet" href="css/common.css" type="text/css" />
+
+<script type="text/javascript">
+  pathId = "StanfordCore";
+  relpath = '';
+</script>
+
+
+  <script type="text/javascript" charset="utf-8" src="js/jquery.js"></script>
+
+  <script type="text/javascript" charset="utf-8" src="js/app.js"></script>
+
+
+  </head>
+  <body>
+    <div class="nav_wrap">
+      <iframe id="nav" src="class_list.html?1"></iframe>
+      <div id="resizer"></div>
+    </div>
+
+    <div id="main" tabindex="-1">
+      <div id="header">
+        <div id="menu">
+  
+    <a href="_index.html">Index (S)</a> &raquo;
+    
+    
+    <span class="title">StanfordCore</span>
+  
+</div>
+
+        <div id="search">
+  
+    <a class="full_list_link" id="class_list_link"
+        href="class_list.html">
+
+        <svg width="24" height="24">
+          <rect x="0" y="4" width="24" height="4" rx="1" ry="1"></rect>
+          <rect x="0" y="12" width="24" height="4" rx="1" ry="1"></rect>
+          <rect x="0" y="20" width="24" height="4" rx="1" ry="1"></rect>
+        </svg>
+    </a>
+  
+</div>
+        <div class="clear"></div>
+      </div>
+
+      <div id="content"><h1>Module: StanfordCore
+  
+  
+  
+</h1>
+<div class="box_info">
+  
+
+  
+  
+  
+  
+  
+
+  
+
+  
+  <dl>
+    <dt>Defined in:</dt>
+    <dd>lib/stanford_core/mention.rb<span class="defines">,<br />
+  lib/stanford_core/text.rb,<br /> lib/stanford_core/tree.rb,<br /> lib/stanford_core/token.rb,<br /> lib/stanford_core/sentence.rb,<br /> lib/stanford_core/coref_chain.rb,<br /> lib/stanford_core/nlp_wrapper.rb,<br /> lib/stanford_core/nlp_pipeline.rb,<br /> lib/stanford_core/coref_chain_map.rb</span>
+</dd>
+  </dl>
+  
+</div>
+
+<h2>Overview</h2><div class="docstring">
+  <div class="discussion">
+    
+<p><a href="https://nlp.stanford.edu/nlp/javadoc/javanlp-3.5.0/edu/stanford/nlp/dcoref/Dictionaries.MentionType.html">nlp.stanford.edu/nlp/javadoc/javanlp-3.5.0/edu/stanford/nlp/dcoref/Dictionaries.MentionType.html</a></p>
+
+
+  </div>
+</div>
+<div class="tags">
+  
+
+</div><h2>Defined Under Namespace</h2>
+<p class="children">
+  
+    
+  
+    
+      <strong class="classes">Classes:</strong> <span class='object_link'><a href="StanfordCore/CorefChain.html" title="StanfordCore::CorefChain (class)">CorefChain</a></span>, <span class='object_link'><a href="StanfordCore/CorefChainMap.html" title="StanfordCore::CorefChainMap (class)">CorefChainMap</a></span>, <span class='object_link'><a href="StanfordCore/Mention.html" title="StanfordCore::Mention (class)">Mention</a></span>, <span class='object_link'><a href="StanfordCore/NlpPipeline.html" title="StanfordCore::NlpPipeline (class)">NlpPipeline</a></span>, <span class='object_link'><a href="StanfordCore/NlpWrapper.html" title="StanfordCore::NlpWrapper (class)">NlpWrapper</a></span>, <span class='object_link'><a href="StanfordCore/Sentence.html" title="StanfordCore::Sentence (class)">Sentence</a></span>, <span class='object_link'><a href="StanfordCore/Text.html" title="StanfordCore::Text (class)">Text</a></span>, <span class='object_link'><a href="StanfordCore/Token.html" title="StanfordCore::Token (class)">Token</a></span>, <span class='object_link'><a href="StanfordCore/Tree.html" title="StanfordCore::Tree (class)">Tree</a></span>
+    
+  
+</p>
+
+
+
+
+
+
+
+
+
+</div>
+
+      <div id="footer">
+  Generated on Tue Jun  9 20:32:31 2020 by
+  <a href="http://yardoc.org" title="Yay! A Ruby Documentation Tool" target="_parent">yard</a>
+  0.9.25 (ruby-2.6.2).
+</div>
+
+    </div>
+  </body>
+</html>
