+++ doc/TextPreprocessor.html
@@ -427,7 +427,7 @@
     <p class="tag_title">Examples:</p>
     
       
-      <pre class="example code"><code>&quot;Open a pull-request and revit it&quot; =&gt; &quot;Open a pull-request and review pull-request&quot;</code></pre>
+      <pre class="example code"><code>&quot;Open a pull-request and review it&quot; =&gt; &quot;Open a pull-request and review pull-request&quot;</code></pre>
     
   </div>
 
@@ -551,7 +551,7 @@
 </div>
 
       <div id="footer">
-  Generated on Tue Jun  9 19:59:00 2020 by
+  Generated on Tue Jun  9 20:32:31 2020 by
   <a href="http://yardoc.org" title="Yay! A Ruby Documentation Tool" target="_parent">yard</a>
   0.9.25 (ruby-2.6.2).
 </div>
