+++ doc/_index.html
@@ -94,12 +94,16 @@
           <ul>
             
               <li>
-                <span class='object_link'><a href="CorefChain.html" title="CorefChain (class)">CorefChain</a></span>
+                <span class='object_link'><a href="StanfordCore/CorefChain.html" title="StanfordCore::CorefChain (class)">CorefChain</a></span>
+                
+                  <small>(StanfordCore)</small>
                 
               </li>
             
               <li>
-                <span class='object_link'><a href="CorefChainMap.html" title="CorefChainMap (class)">CorefChainMap</a></span>
+                <span class='object_link'><a href="StanfordCore/CorefChainMap.html" title="StanfordCore::CorefChainMap (class)">CorefChainMap</a></span>
+                
+                  <small>(StanfordCore)</small>
                 
               </li>
             
@@ -112,7 +116,9 @@
           <ul>
             
               <li>
-                <span class='object_link'><a href="Mention.html" title="Mention (class)">Mention</a></span>
+                <span class='object_link'><a href="StanfordCore/Mention.html" title="StanfordCore::Mention (class)">Mention</a></span>
+                
+                  <small>(StanfordCore)</small>
                 
               </li>
             
@@ -125,12 +131,16 @@
           <ul>
             
               <li>
-                <span class='object_link'><a href="NlpPipeline.html" title="NlpPipeline (class)">NlpPipeline</a></span>
+                <span class='object_link'><a href="StanfordCore/NlpPipeline.html" title="StanfordCore::NlpPipeline (class)">NlpPipeline</a></span>
+                
+                  <small>(StanfordCore)</small>
                 
               </li>
             
               <li>
-                <span class='object_link'><a href="NlpWrapper.html" title="NlpWrapper (class)">NlpWrapper</a></span>
+                <span class='object_link'><a href="StanfordCore/NlpWrapper.html" title="StanfordCore::NlpWrapper (class)">NlpWrapper</a></span>
+                
+                  <small>(StanfordCore)</small>
                 
               </li>
             
@@ -143,7 +153,14 @@
           <ul>
             
               <li>
-                <span class='object_link'><a href="Sentence.html" title="Sentence (class)">Sentence</a></span>
+                <span class='object_link'><a href="StanfordCore/Sentence.html" title="StanfordCore::Sentence (class)">Sentence</a></span>
+                
+                  <small>(StanfordCore)</small>
+                
+              </li>
+            
+              <li>
+                <span class='object_link'><a href="StanfordCore.html" title="StanfordCore (module)">StanfordCore</a></span>
                 
               </li>
             
@@ -156,7 +173,9 @@
           <ul>
             
               <li>
-                <span class='object_link'><a href="Text.html" title="Text (class)">Text</a></span>
+                <span class='object_link'><a href="StanfordCore/Text.html" title="StanfordCore::Text (class)">Text</a></span>
+                
+                  <small>(StanfordCore)</small>
                 
               </li>
             
@@ -166,12 +185,16 @@
               </li>
             
               <li>
-                <span class='object_link'><a href="Token.html" title="Token (class)">Token</a></span>
+                <span class='object_link'><a href="StanfordCore/Token.html" title="StanfordCore::Token (class)">Token</a></span>
+                
+                  <small>(StanfordCore)</small>
                 
               </li>
             
               <li>
-                <span class='object_link'><a href="Tree.html" title="Tree (class)">Tree</a></span>
+                <span class='object_link'><a href="StanfordCore/Tree.html" title="StanfordCore::Tree (class)">Tree</a></span>
+                
+                  <small>(StanfordCore)</small>
                 
               </li>
             
@@ -192,7 +215,7 @@
 </div>
 
       <div id="footer">
-  Generated on Tue Jun  9 19:59:00 2020 by
+  Generated on Tue Jun  9 20:32:30 2020 by
   <a href="http://yardoc.org" title="Yay! A Ruby Documentation Tool" target="_parent">yard</a>
   0.9.25 (ruby-2.6.2).
 </div>
