+++ doc/class_list.html
@@ -43,7 +43,7 @@
 
       <ul id="full_list" class="class">
         <li id="object_" class="odd"><div class="item" style="padding-left:30px"><span class='object_link'><a href="top-level-namespace.html" title="Top Level Namespace (root)">Top Level Namespace</a></span></div></li>
-<li id='object_ActivityPhraseBuilder' class='even'><div class='item' style='padding-left:30px'><span class='object_link'><a href="ActivityPhraseBuilder.html" title="ActivityPhraseBuilder (class)">ActivityPhraseBuilder</a></span> &lt; Object<small class='search_info'>Top Level Namespace</small></div></li><li id='object_CorefChain' class='odd'><div class='item' style='padding-left:30px'><span class='object_link'><a href="CorefChain.html" title="CorefChain (class)">CorefChain</a></span> &lt; NlpWrapper<small class='search_info'>Top Level Namespace</small></div></li><li id='object_CorefChainMap' class='even'><div class='item' style='padding-left:30px'><span class='object_link'><a href="CorefChainMap.html" title="CorefChainMap (class)">CorefChainMap</a></span> &lt; NlpWrapper<small class='search_info'>Top Level Namespace</small></div></li><li id='object_Mention' class='odd'><div class='item' style='padding-left:30px'><span class='object_link'><a href="Mention.html" title="Mention (class)">Mention</a></span> &lt; NlpWrapper<small class='search_info'>Top Level Namespace</small></div></li><li id='object_NlpPipeline' class='even'><div class='item' style='padding-left:30px'><span class='object_link'><a href="NlpPipeline.html" title="NlpPipeline (class)">NlpPipeline</a></span> &lt; Object<small class='search_info'>Top Level Namespace</small></div></li><li id='object_NlpWrapper' class='odd'><div class='item' style='padding-left:30px'><span class='object_link'><a href="NlpWrapper.html" title="NlpWrapper (class)">NlpWrapper</a></span> &lt; Object<small class='search_info'>Top Level Namespace</small></div></li><li id='object_Sentence' class='even'><div class='item' style='padding-left:30px'><span class='object_link'><a href="Sentence.html" title="Sentence (class)">Sentence</a></span> &lt; Object<small class='search_info'>Top Level Namespace</small></div></li><li id='object_Text' class='odd'><div class='item' style='padding-left:30px'><span class='object_link'><a href="Text.html" title="Text (class)">Text</a></span> &lt; Object<small class='search_info'>Top Level Namespace</small></div></li><li id='object_TextPreprocessor' class='even'><div class='item' style='padding-left:30px'><span class='object_link'><a href="TextPreprocessor.html" title="TextPreprocessor (class)">TextPreprocessor</a></span> &lt; Object<small class='search_info'>Top Level Namespace</small></div></li><li id='object_Token' class='odd'><div class='item' style='padding-left:30px'><span class='object_link'><a href="Token.html" title="Token (class)">Token</a></span> &lt; NlpWrapper<small class='search_info'>Top Level Namespace</small></div></li><li id='object_Tree' class='even'><div class='item' style='padding-left:30px'><span class='object_link'><a href="Tree.html" title="Tree (class)">Tree</a></span> &lt; Object<small class='search_info'>Top Level Namespace</small></div></li><li id='object_TreeActivitiesIdentifier' class='odd'><div class='item' style='padding-left:30px'><span class='object_link'><a href="TreeActivitiesIdentifier.html" title="TreeActivitiesIdentifier (class)">TreeActivitiesIdentifier</a></span> &lt; Object<small class='search_info'>Top Level Namespace</small></div></li>
+<li id='object_ActivityPhraseBuilder' class='even'><div class='item' style='padding-left:30px'><span class='object_link'><a href="ActivityPhraseBuilder.html" title="ActivityPhraseBuilder (class)">ActivityPhraseBuilder</a></span> &lt; Object<small class='search_info'>Top Level Namespace</small></div></li><li id='object_StanfordCore' class='odd'><div class='item' style='padding-left:30px'><a class='toggle'></a> <span class='object_link'><a href="StanfordCore.html" title="StanfordCore (module)">StanfordCore</a></span><small class='search_info'>Top Level Namespace</small></div><ul><li id='object_StanfordCore::CorefChain' class='collapsed even'><div class='item' style='padding-left:45px'><span class='object_link'><a href="StanfordCore/CorefChain.html" title="StanfordCore::CorefChain (class)">CorefChain</a></span> &lt; NlpWrapper<small class='search_info'>StanfordCore</small></div></li><li id='object_StanfordCore::CorefChainMap' class='collapsed odd'><div class='item' style='padding-left:45px'><span class='object_link'><a href="StanfordCore/CorefChainMap.html" title="StanfordCore::CorefChainMap (class)">CorefChainMap</a></span> &lt; NlpWrapper<small class='search_info'>StanfordCore</small></div></li><li id='object_StanfordCore::Mention' class='collapsed even'><div class='item' style='padding-left:45px'><span class='object_link'><a href="StanfordCore/Mention.html" title="StanfordCore::Mention (class)">Mention</a></span> &lt; NlpWrapper<small class='search_info'>StanfordCore</small></div></li><li id='object_StanfordCore::NlpPipeline' class='collapsed odd'><div class='item' style='padding-left:45px'><span class='object_link'><a href="StanfordCore/NlpPipeline.html" title="StanfordCore::NlpPipeline (class)">NlpPipeline</a></span> &lt; Object<small class='search_info'>StanfordCore</small></div></li><li id='object_StanfordCore::NlpWrapper' class='collapsed even'><div class='item' style='padding-left:45px'><span class='object_link'><a href="StanfordCore/NlpWrapper.html" title="StanfordCore::NlpWrapper (class)">NlpWrapper</a></span> &lt; Object<small class='search_info'>StanfordCore</small></div></li><li id='object_StanfordCore::Sentence' class='collapsed odd'><div class='item' style='padding-left:45px'><span class='object_link'><a href="StanfordCore/Sentence.html" title="StanfordCore::Sentence (class)">Sentence</a></span> &lt; Object<small class='search_info'>StanfordCore</small></div></li><li id='object_StanfordCore::Text' class='collapsed even'><div class='item' style='padding-left:45px'><span class='object_link'><a href="StanfordCore/Text.html" title="StanfordCore::Text (class)">Text</a></span> &lt; Object<small class='search_info'>StanfordCore</small></div></li><li id='object_StanfordCore::Token' class='collapsed odd'><div class='item' style='padding-left:45px'><span class='object_link'><a href="StanfordCore/Token.html" title="StanfordCore::Token (class)">Token</a></span> &lt; NlpWrapper<small class='search_info'>StanfordCore</small></div></li><li id='object_StanfordCore::Tree' class='collapsed even'><div class='item' style='padding-left:45px'><span class='object_link'><a href="StanfordCore/Tree.html" title="StanfordCore::Tree (class)">Tree</a></span> &lt; Object<small class='search_info'>StanfordCore</small></div></li></ul></li><li id='object_TextPreprocessor' class='odd'><div class='item' style='padding-left:30px'><span class='object_link'><a href="TextPreprocessor.html" title="TextPreprocessor (class)">TextPreprocessor</a></span> &lt; Object<small class='search_info'>Top Level Namespace</small></div></li><li id='object_TreeActivitiesIdentifier' class='even'><div class='item' style='padding-left:30px'><span class='object_link'><a href="TreeActivitiesIdentifier.html" title="TreeActivitiesIdentifier (class)">TreeActivitiesIdentifier</a></span> &lt; Object<small class='search_info'>Top Level Namespace</small></div></li>
 
       </ul>
     </div>
