+++ doc/file.README.html
@@ -63,7 +63,7 @@
 </div></div>
 
       <div id="footer">
-  Generated on Tue Jun  9 19:59:00 2020 by
+  Generated on Tue Jun  9 20:32:31 2020 by
   <a href="http://yardoc.org" title="Yay! A Ruby Documentation Tool" target="_parent">yard</a>
   0.9.25 (ruby-2.6.2).
 </div>
