+++ doc/top-level-namespace.html
@@ -82,9 +82,11 @@
 <p class="children">
   
     
+      <strong class="modules">Modules:</strong> <span class='object_link'><a href="StanfordCore.html" title="StanfordCore (module)">StanfordCore</a></span>
+    
   
     
-      <strong class="classes">Classes:</strong> <span class='object_link'><a href="ActivityPhraseBuilder.html" title="ActivityPhraseBuilder (class)">ActivityPhraseBuilder</a></span>, <span class='object_link'><a href="CorefChain.html" title="CorefChain (class)">CorefChain</a></span>, <span class='object_link'><a href="CorefChainMap.html" title="CorefChainMap (class)">CorefChainMap</a></span>, <span class='object_link'><a href="Mention.html" title="Mention (class)">Mention</a></span>, <span class='object_link'><a href="NlpPipeline.html" title="NlpPipeline (class)">NlpPipeline</a></span>, <span class='object_link'><a href="NlpWrapper.html" title="NlpWrapper (class)">NlpWrapper</a></span>, <span class='object_link'><a href="Sentence.html" title="Sentence (class)">Sentence</a></span>, <span class='object_link'><a href="Text.html" title="Text (class)">Text</a></span>, <span class='object_link'><a href="TextPreprocessor.html" title="TextPreprocessor (class)">TextPreprocessor</a></span>, <span class='object_link'><a href="Token.html" title="Token (class)">Token</a></span>, <span class='object_link'><a href="Tree.html" title="Tree (class)">Tree</a></span>, <span class='object_link'><a href="TreeActivitiesIdentifier.html" title="TreeActivitiesIdentifier (class)">TreeActivitiesIdentifier</a></span>
+      <strong class="classes">Classes:</strong> <span class='object_link'><a href="ActivityPhraseBuilder.html" title="ActivityPhraseBuilder (class)">ActivityPhraseBuilder</a></span>, <span class='object_link'><a href="TextPreprocessor.html" title="TextPreprocessor (class)">TextPreprocessor</a></span>, <span class='object_link'><a href="TreeActivitiesIdentifier.html" title="TreeActivitiesIdentifier (class)">TreeActivitiesIdentifier</a></span>
     
   
 </p>
@@ -100,7 +102,7 @@
 </div>
 
       <div id="footer">
-  Generated on Tue Jun  9 19:59:00 2020 by
+  Generated on Tue Jun  9 20:32:31 2020 by
   <a href="http://yardoc.org" title="Yay! A Ruby Documentation Tool" target="_parent">yard</a>
   0.9.25 (ruby-2.6.2).
 </div>
