+++ spec/lib/text_preprocessor_spec.rb
@@ -3,28 +3,34 @@
 require 'rails_helper'
 
 RSpec.describe TextPreprocessor do
-  let(:text) do
-    file_fixture('text.txt').read
-  end
-  let(:subject) { described_class.new(Text.new(text)) }
-  
-  describe ".substitute_coreferences" do
+  let(:text) { file_fixture('text.txt').read }
+  let(:preprocessed_text) { file_fixture('preprocessed_text.txt').read }
+  let(:subject) { described_class.new(text) }
+
+  describe '.substitute_coreferences' do
     let(:expected_text) do
-      'Alessandro, please review pull-request so that we can get pull-request to QA. Sure. Alessandro will review pull-request today. Before reviewing pull-request, let all tests pass.'
+      'Alessandro, please review this this pull-request so that we can get this pull-request to QA. Sure. I will review this pull-request today. Before reviewing this pull-request, let all tests pass.'
     end
 
     it 'works' do
-      expect(subject.substitute_coreferences).to eq(expected_text)
+      expect(subject.substitute_coreferences.to_s).to eq(expected_text)
     end
   end
 
-  describe ".remove_teterminers" do
+  describe '.remove_teterminers' do
     let(:expected_text) do
-      'Alessandro, please review pull-request so that we can get it to QA. Sure. Alessandro will review it today. Before reviewing it, let tests pass.'
+      'Alessandro, please review pull-request so that we can get it to QA. Sure. I will review it today. Before reviewing it, let tests pass.'
     end
 
     it 'works' do
-      expect(subject.remove_determiners).to eq(expected_text)
+      expect(subject.remove_determiners.to_s).to eq(expected_text)
+    end
+  end
+
+  describe 'fully processed text' do
+    it 'works as expected' do
+      expect(subject.substitute_coreferences.remove_determiners.to_s)
+        .to eq(preprocessed_text)
     end
   end
 end
